+++ src/components/background.tsx
@@ -1,7 +1,9 @@
+'use dom';
+
 import styles from '@/styles/background.module.css';
 import classNames from 'classnames';
 
-export default function Background() {
+export default function Background(_: { dom?: import('expo/dom').DOMProps }) {
   return (
     <div className="z-[-1] bg-[#10141A] fixed top-0 left-0 bottom-0 right-0 pointer-events-none pt-24 pb-32 px-4 md:px-0 flex justify-center">
       <div
